apps/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.21.5
pandas==1.3.5
scikit_learn==1.0.2
xgboost==1.6.1
